// src/app/model/event.ts
export class Event {
    id?: string;
    title?: string;
    dateStart?: Date;
    dateEnd?: Date;
    categoryName?: string;
    categoryID?: string;
    address?: string;
    image?: string;
    note?: string;
    userCreate?: string;
    dateCreate?: Date;
    userUpdate?: string;
    dateUpdate?: Date;
    messWarning?: string;
}