// src/app/model/eventDetail.ts
export class EventDetail {
    id?: string;
    eventID?: string;
    title?: string;
    dateStart?: Date;
    dateEnd?: Date;
    address?: string;
    file?: string;
    note?: string;
    userCreate?: string;
    dateCreate?: Date;
    userUpdate?: string;
    dateUpdate?: Date;
    messWarning?: string;
}