// src/app/model/org-structure.ts
export class OrgStructure {
    id?: string;
    code?: string;
    orgStructureName?: string;
    headOrgStructureID?: string;
    headOrgStructureName?: string;
    note?: string;
    userCreate?: string;
    dateCreate?: Date;
    userUpdate?: string;
    dateUpdate?: Date;
    messWarning?: string;
}

export class OrgStructureMulti {
    value = "";
    label = "";
}